+++ src/mlpack/methods/hmm/hmm_generate_main.cpp
@@ -30,12 +30,17 @@ using namespace mlpack::math;
 using namespace arma;
 using namespace std;
 
-PROGRAM_INFO("Hidden Markov Model (HMM) Sequence Generator",
-    // Short description.
+// Program Name.
+BINDING_NAME("Hidden Markov Model (HMM) Sequence Generator");
+
+// Short description.
+BINDING_SHORT_DESC(
     "A utility to generate random sequences from a pre-trained Hidden Markov "
     "Model (HMM).  The length of the desired sequence can be specified, and a "
-    "random sequence of observations is returned.",
-    // Long description.
+    "random sequence of observations is returned.");
+
+// Long description.
+BINDING_LONG_DESC(
     "This utility takes an already-trained HMM, specified as the " +
     PRINT_PARAM_STRING("model") + " parameter, and generates a random "
     "observation sequence and hidden state sequence based on its parameters. "
@@ -45,22 +50,26 @@ PROGRAM_INFO("Hidden Markov Model (HMM) Sequence Generator",
     " parameter."
     "\n\n"
     "The state to start the sequence in may be specified with the " +
-    PRINT_PARAM_STRING("start_state") + " parameter."
-    "\n\n"
+    PRINT_PARAM_STRING("start_state") + " parameter.");
+
+// Example.
+BINDING_EXAMPLE(
     "For example, to generate a sequence of length 150 from the HMM " +
     PRINT_MODEL("hmm") + " and save the observation sequence to " +
     PRINT_DATASET("observations") + " and the hidden state sequence to " +
     PRINT_DATASET("states") + ", the following command may be used: "
     "\n\n" +
     PRINT_CALL("hmm_generate", "model", "hmm", "length", 150, "output",
-        "observations", "state", "states"),
-    SEE_ALSO("@hmm_train", "#hmm_train"),
-    SEE_ALSO("@hmm_loglik", "#hmm_loglik"),
-    SEE_ALSO("@hmm_viterbi", "#hmm_viterbi"),
-    SEE_ALSO("Hidden Mixture Models on Wikipedia",
-        "https://en.wikipedia.org/wiki/Hidden_Markov_model"),
-    SEE_ALSO("mlpack::hmm::HMM class documentation",
-        "@doxygen/classmlpack_1_1hmm_1_1HMM.html"));
+        "observations", "state", "states"));
+
+// See also...
+BINDING_SEE_ALSO("@hmm_train", "#hmm_train");
+BINDING_SEE_ALSO("@hmm_loglik", "#hmm_loglik");
+BINDING_SEE_ALSO("@hmm_viterbi", "#hmm_viterbi");
+BINDING_SEE_ALSO("Hidden Mixture Models on Wikipedia",
+        "https://en.wikipedia.org/wiki/Hidden_Markov_model");
+BINDING_SEE_ALSO("mlpack::hmm::HMM class documentation",
+        "@doxygen/classmlpack_1_1hmm_1_1HMM.html");
 
 PARAM_MODEL_IN_REQ(HMMModel, "model", "Trained HMM to generate sequences with.",
     "m");