+++ src/mlpack/methods/kernel_pca/kernel_pca_main.cpp
@@ -40,12 +40,17 @@ using namespace mlpack::util;
 using namespace std;
 using namespace arma;
 
-PROGRAM_INFO("Kernel Principal Components Analysis",
-    // Short description.
+// Program Name.
+BINDING_NAME("Kernel Principal Components Analysis");
+
+// Short description.
+BINDING_SHORT_DESC(
     "An implementation of Kernel Principal Components Analysis (KPCA).  This "
     "can be used to perform nonlinear dimensionality reduction or preprocessing"
-    " on a given dataset.",
-    // Long description.
+    " on a given dataset.");
+
+// Long description.
+BINDING_LONG_DESC(
     "This program performs Kernel Principal Components Analysis (KPCA) on the "
     "specified dataset with the specified kernel.  This will transform the "
     "data onto the kernel principal components, and optionally reduce the "
@@ -54,13 +59,6 @@ PROGRAM_INFO("Kernel Principal Components Analysis",
     "\n\n"
     "For the case where a linear kernel is used, this reduces to regular "
     "PCA."
-    "\n\n"
-    "For example, the following command will perform KPCA on the dataset " +
-    PRINT_DATASET("input") + " using the Gaussian kernel, and saving the "
-    "transformed data to " + PRINT_DATASET("transformed") + ": "
-    "\n\n" +
-    PRINT_CALL("kernel_pca", "input", "input", "kernel", "gaussian", "output",
-        "transformed") +
     "\n\n"
     "The kernels that are supported are listed below:"
     "\n\n"
@@ -98,13 +96,24 @@ PROGRAM_INFO("Kernel Principal Components Analysis",
     "the kernel matrix; to specify the sampling scheme, the " +
     PRINT_PARAM_STRING("sampling") + " parameter is used.  The "
     "sampling scheme for the Nystroem method can be chosen from the "
-    "following list: 'kmeans', 'random', 'ordered'.",
-    SEE_ALSO("Kernel principal component analysis on Wikipedia",
-        "https://en.wikipedia.org/wiki/Kernel_principal_component_analysis"),
-    SEE_ALSO("Kernel Principal Component Analysis (pdf)",
-        "http://pca.narod.ru/scholkopf_kernel.pdf"),
-    SEE_ALSO("mlpack::kpca::KernelPCA class documentation",
-        "@doxygen/classmlpack_1_1kpca_1_1KernelPCA.html"));
+    "following list: 'kmeans', 'random', 'ordered'.");
+
+// Example.
+BINDING_EXAMPLE(
+    "For example, the following command will perform KPCA on the dataset " +
+    PRINT_DATASET("input") + " using the Gaussian kernel, and saving the "
+    "transformed data to " + PRINT_DATASET("transformed") + ": "
+    "\n\n" +
+    PRINT_CALL("kernel_pca", "input", "input", "kernel", "gaussian", "output",
+        "transformed"));
+
+// See also...
+BINDING_SEE_ALSO("Kernel principal component analysis on Wikipedia",
+        "https://en.wikipedia.org/wiki/Kernel_principal_component_analysis");
+BINDING_SEE_ALSO("Kernel Principal Component Analysis (pdf)",
+        "http://pca.narod.ru/scholkopf_kernel.pdf");
+BINDING_SEE_ALSO("mlpack::kpca::KernelPCA class documentation",
+        "@doxygen/classmlpack_1_1kpca_1_1KernelPCA.html");
 
 PARAM_MATRIX_IN_REQ("input", "Input dataset to perform KPCA on.", "i");
 PARAM_MATRIX_OUT("output", "Matrix to save modified dataset to.", "o");