+++ src/mlpack/tests/rbm_network_test.cpp
@@ -88,14 +88,13 @@ BOOST_AUTO_TEST_CASE(BinaryRBMClassificationTest)
 
   for (size_t i = 0; i < trainData.n_cols; ++i)
   {
-    model.HiddenMean(std::move(trainData.col(i)), std::move(output));
+    model.HiddenMean(trainData.col(i), output);
     XRbm.col(i) = output;
   }
 
   for (size_t i = 0; i < testData.n_cols; ++i)
   {
-    model.HiddenMean(std::move(testData.col(i)),
-      std::move(output));
+    model.HiddenMean(testData.col(i), output);
     YRbm.col(i) = output;
   }
   const size_t numClasses = 10; // Number of classes.
@@ -189,15 +188,13 @@ BOOST_AUTO_TEST_CASE(ssRBMClassificationTest)
 
   for (size_t i = 0; i < trainData.n_cols; ++i)
   {
-    modelssRBM.HiddenMean(std::move(trainData.col(i)),
-        std::move(output));
+    modelssRBM.HiddenMean(trainData.col(i), output);
     XRbm.col(i) = output;
   }
 
   for (size_t i = 0; i < testData.n_cols; ++i)
   {
-    modelssRBM.HiddenMean(std::move(testData.col(i)),
-      std::move(output));
+    modelssRBM.HiddenMean(testData.col(i), output);
     YRbm.col(i) = output;
   }
   const size_t numClasses = 10; // Number of classes.
@@ -238,7 +235,7 @@ void BuildVanillaNetwork(MatType& trainData,
   arma::vec calculatedFreeEnergy(4, arma::fill::zeros);
   for (size_t i = 0; i < trainData.n_cols; ++i)
   {
-    calculatedFreeEnergy(i) = model.FreeEnergy(std::move(trainData.col(i)));
+    calculatedFreeEnergy(i) = model.FreeEnergy(trainData.col(i));
   }
 
   for (size_t i = 0; i < freeEnergy.n_elem; ++i)